+++ packages/sdk/src/tokens/token-addresses.ts
@@ -201,7 +201,6 @@ export const EVM_LBTC_ADDRESSES: TokenAddressesPerEnv<Token.LBTC, ChainId> = {
     [ChainId.base]: '0xecAc9C5F704e954931349Da37F60E39f515c11c1',
     [ChainId.berachain]: '0xecAc9C5F704e954931349Da37F60E39f515c11c1',
     [ChainId.binanceSmartChain]: '0xecAc9C5F704e954931349Da37F60E39f515c11c1',
-    [ChainId.corn]: '0xecAc9C5F704e954931349Da37F60E39f515c11c1',
     [ChainId.etherlink]: '0xecAc9C5F704e954931349Da37F60E39f515c11c1',
     [ChainId.katana]: '0xecAc9C5F704e954931349Da37F60E39f515c11c1',
     [ChainId.megaeth]: '0xecAc9C5F704e954931349Da37F60E39f515c11c1',
@@ -278,11 +277,6 @@ const EVM_SUPPORTING_TOKEN_ADDRESSES: TokenAddresses<SupportingToken, ChainId> =
     [Token.wBTC]: {
       [Env.prod]: {
         [ChainId.ethereum]: '0x2260fac5e5542a773aa44fbcfedf7c193bc2c599',
-      },
-    },
-    [Token.wBTCN]: {
-      [Env.prod]: {
-        [ChainId.corn]: '0xda5dDd7270381A7C2717aD10D1c0ecB19e3CDFb2',
       },
     },
   };